pyproject.toml
[project]
name = "pyoverkiz"
version = "2.0.0"
description = "A fully asynchronous API client for interaction with smart devices connected to OverKiz, supporting multiple vendors such as Somfy TaHoma and Atlantic Cozytouch."
readme = "README.md"
authors = [
    {name = "Mick Vleeshouwer", email = "mick@imick.nl"},
    {name = "Vincent Le Bourlot"},
    {name = "Thibaut Etienne"},
]
license = {text = "MIT"}
requires-python = "<4.0,>=3.12"
packages = [
    { include = "pyoverkiz" }
]
dependencies = [
    "aiohttp<4.0.0,>=3.10.3",
    "pyhumps<4.0.0,>=3.8.0",
    "backoff<3.0,>=1.10.0",
    "attrs>=21.2",
    "cattrs>=24.1.0",
    "boto3<2.0.0,>=1.18.59",
    "warrant-lite<2.0.0,>=1.0.4",
]

[project.optional-dependencies]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-autorefs>=1.0.0",
    "pymdown-extensions>=10.0",
]

[project.urls]
homepage = "https://github.com/iMicknl/python-overkiz-api"
repository = "https://github.com/iMicknl/python-overkiz-api"

[dependency-groups]
dev = [
    "pytest>=8.3.2,<10.0.0",
    "pytest-cov>=5,<8",
    "pytest-asyncio>=0.23.8,<1.4.0",
    "ruff>=0.12.0",
    "mypy>=1.16.1",
    "ty>=0.0.8",
    "prek>=0.2.27",
]

[tool.ruff.lint]
select = [
    # pydocstyle
    "D",
    # flake8-async
    "ASYNC",
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # ruff
    "RUF",
    # flake8-bandit (security)
    "S",
    # flake8-print
    "T",
    # flake8-comprehensions
    "C4",
]
ignore = ["E501"] # Line too long

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # Allow assert in tests

[tool.ruff.lint.pydocstyle]
convention = "google"  # Accepts: "google", "numpy", or "pep257".

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
  "pyoverkiz/*",
]

[tool.hatch.build.targets.wheel]
packages = ["pyoverkiz"]
only-packages = true

[tool.mypy]
exclude = ["tests/"]
ignore_missing_imports = true

[tool.ty.src]
exclude = ["tests/"]
